refactor(render/sarif): split into submodule directory (#57 3/3)
src/render/sarif.rs (1205 LOC) -> src/render/sarif/:
- mod.rs (73): docstring, consts, module decls, pub use render
- document.rs (32): pub fn render orchestrator
- rules.rs (128): fn rules + fn rule
- results.rs (350): fn results (twelve-loop finding emitter)
- helpers.rs (58): fingerprint, plugin_sarif_level, synthetic_location, sarif_level
- tests.rs (588): #[cfg(test)] mod tests

Biggest non-test file: results.rs at 350 LOC. Pure cut-paste, zero
behavior change. Public API path crate::render::sarif::render preserved.
429 lib tests pass, clippy clean.

Closes #57 (all three splits landed: run.rs #62, maintainer.rs #64, sarif.rs).
